fix: type assertion bugs feat: update readme feat: add printing of opcodes to provide a starting point for the user
changed the Value struct member Value to be an interface type. This enables the program to type assert the value each time we execute an opcode. If the type is not a pointer to a big.Int, then execution stops
updated readme to help the user with the onboarding process
print each opcode for debugging purposes.